backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.13
pydantic[email]==1.10.13

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# NLP and Text Processing
spacy==3.7.2
nltk==3.8.1
langdetect==1.0.9
textblob==0.17.1
emoji==2.9.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
python-dateutil==2.8.2

# File handling
aiofiles==23.2.1
python-magic==0.4.27
chardet==5.2.0

# API utilities
httpx==0.26.0
python-dotenv==1.0.0
email-validator==2.1.0

# Background tasks
celery==5.3.4
flower==2.0.1

# Storage
boto3==1.34.14  # For S3 support
minio==7.2.2    # Alternative S3-compatible storage

# Export formats
reportlab==4.0.8  # PDF generation
openpyxl==3.1.2   # Excel export
jinja2==3.1.3     # Template rendering

# Monitoring and logging
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk==1.39.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # For API testing
factory-boy==3.3.0
faker==22.0.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
ipython==8.19.0